refactor(ast): name the engine defineMacro.ts and split presets per file
Rename the macro engine to src/defineMacro.ts at the package root (it holds
defineMacro, composeMacros, applyMacros, and the Macro type). The macros
folder now has one preset per file (macroDiscriminatorEnum, macroSimplifyUnion,
macroEnumName), re-exported from macros/index.ts on the @kubb/ast/macros subpath.
