+++ packages/ast/src/macros/index.ts
@@ -1,3 +1,5 @@
-// Built-in macro presets. The macro engine (`defineMacro`, `composeMacros`, `applyMacros`, and the
-// `Macro` type) lives on the `@kubb/ast` root barrel.
-export { macroDiscriminatorEnum, macroEnumName, macroSimplifyUnion } from './presets.ts'
+// Built-in macro presets, one per file. The macro engine (`defineMacro`, `composeMacros`,
+// `applyMacros`, and the `Macro` type) lives on the `@kubb/ast` root barrel.
+export { macroDiscriminatorEnum } from './macroDiscriminatorEnum.ts'
+export { macroEnumName } from './macroEnumName.ts'
+export { macroSimplifyUnion } from './macroSimplifyUnion.ts'
